updated tools

+++ tools/compress_data_grid.py
@@ -91,7 +91,7 @@ for nSnap in range(nSnapshots):
       data_all[ procStart_z:procEnd_z, procStart_y:procEnd_y, procStart_x:procEnd_x] = data_local
       inFile.close()
     if key=='grav_density': print '  {0}   {1}   {2}'.format(data_all.mean(), data_all.min(), data_all.max())
-    fileSnap.create_dataset( key, data=data_all.astype(np.float32) )
+    fileSnap.create_dataset( key, data=data_all )
     fileSnap.attrs['max_'+ key ] = data_all.max()
     fileSnap.attrs['min_'+ key ] = data_all.min()
     fileSnap.attrs['mean_'+ key ] = data_all.mean()
